change the riseAngle and zenithAngle from BigDecimal to float. Add a modulus function to riseAngle so that it can cycle from 0 to 1 and then back to 0. update unit tests

+++ simLifeApplication/src/main/java/barsuift/simLife/j2d/panel/SunBrightnessPanel.java
@@ -86,7 +86,7 @@ public class SunBrightnessPanel extends JPanel implements ChangeListener, Subscr
     @Override
     public void stateChanged(ChangeEvent e) {
         JSlider source = (JSlider) e.getSource();
-        int brightness = (int) source.getValue();
+        int brightness = source.getValue();
         sun.setBrightness(PercentHelper.getDecimalValue(brightness));
     }
